Better install script, skips node or redis if present

+++ install.sh
@@ -11,6 +11,8 @@
 #
 ##############################################################
 
+benvenutodir=$PWD
+
 # Update System
 echo 'System Update'
 apt-get -y update
@@ -19,36 +21,44 @@ echo 'Update completed'
 # Install help app
 apt-get -y install libssl-dev git-core pkg-config build-essential curl gcc g++ checkinstall
 
-# Download & Unpack Node.js - v. 0.8.16
-echo 'Download Node.js - v. 0.8.16'
-mkdir /tmp/node-install
-cd /tmp/node-install
-wget http://nodejs.org/dist/v0.8.16/node-v0.8.16.tar.gz
-tar -zxf node-v0.8.16.tar.gz
-echo 'Node.js download & unpack completed'
+# Download & Unpack Node.js - v. 0.8.20 if not installed
+which node
+exists=$?
+if [[ $exists = 1 ]] ; then
+		echo 'Download Node.js - v. 0.8.20'
+		mkdir /tmp/node-install
+		cd /tmp/node-install
+		wget http://nodejs.org/dist/v0.8.20/node-v0.8.20.tar.gz
+		tar -zxf node-v0.8.20.tar.gz
+		echo 'Node.js download & unpack completed'
 
-# Install Node.js
-echo 'Install Node.js'
-cd node-v0.8.16
-./configure && make && checkinstall --install=yes --pkgname=nodejs --pkgversion "0.8.16" --default
-echo 'Node.js install completed'
+		# Install Node.js
+		echo 'Install Node.js'
+		cd node-v0.8.20
+		./configure && make && checkinstall --install=yes --pkgname=nodejs --pkgversion "0.8.20" --default
+		echo 'Node.js install completed'
+fi
 
-# Install Redis
-echo 'Install Redis'
-cd /tmp
-wget http://download.redis.io/redis-stable.tar.gz
-tar xvzf redis-stable.tar.gz
-cd redis-stable
-make
-cd src
-sudo cp redis-server /usr/local/bin/
-sudo cp redis-cli /usr/local/bin/
-echo 'Redis install completed. '
+# Install Redis, only if not installed
+which redis-cli
+exists=$?
+if [[ $exists = 1 ]] ; then
+		echo 'Installing Redis'
+		cd /tmp
+		wget http://download.redis.io/redis-stable.tar.gz
+		tar xvzf redis-stable.tar.gz
+		cd redis-stable
+		make
+		cd src
+		sudo cp redis-server /usr/local/bin/
+		sudo cp redis-cli /usr/local/bin/
+		echo 'Redis install completed. '
+fi
 
 # Install Benvenuto Node
 echo 'Installing Benvenuto Node dependencies'
-cd ~/benvenuto-node/
+cd "$benvenutodir"
 npm install
 npm install -g nodemon
 npm install -g coffee-script
-echo 'Benvenuto Node install successful. Run start.sh to start the server'
+echo 'Benvenuto Node install successful. Run sudo benvenuto.sh to start the server'
